move user function in service file
