+++ routes/users.js
@@ -30,160 +30,9 @@ let day_time = 1000 * 60 * 60 * 24;
 let duration_list = conf.get_conf().duration;
 //const runningEnv = process.env.NODE_ENV || 'prod';
 
-function get_group_home (user) {
-    let group_path = CONFIG.general.home+'/'+user.group;
-    if(user.maingroup!='' && user.maingroup!=null) {
-        group_path = CONFIG.general.home+'/'+user.maingroup+'/'+user.group;
-    }
-    return group_path.replace(/\/+/g, '/');
-}
+const grpsrv = require('../core/groups.service.js')
+const usrsrv = require('../core/users.service.js')
 
-function get_user_home (user) {
-    // todo check  or not if user.uid exist
-    let user_home = CONFIG.general.home + '/' + user.uid;
-    if(CONFIG.general.use_group_in_path) {
-        user_home = get_group_home(user) + '/' + user.uid;
-    }
-    return user_home.replace(/\/+/g, '/');
-}
-
-
-router.user_home = function (user) {
-    return get_user_home(user);
-};
-
-
-var create_group = async function(group_name, owner_name){
-    let mingid = await utils.getGroupAvailableId();
-    let fid = new Date().getTime();
-    let group = {name: group_name, gid: mingid, owner: owner_name};
-    try {
-        await utils.mongo_groups().insertOne(group);
-        await goldap.add_group(group, fid);
-    }
-    catch(e) {
-        logger.error(e);
-        throw 'group creation failed';
-    }
-    try {
-        let created_file = await filer.user_add_group(group, fid);
-        logger.info('File Created: ', created_file);
-    } catch(error){
-        logger.error('Create Group Failed for: ' + group.name, error);
-        throw 'group creation failed';
-    }
-
-    await utils.mongo_events().insertOne({'owner': owner_name, 'date': new Date().getTime(), 'action': 'create group ' + group_name , 'logs': [group_name + '.' + fid + '.update']});
-
-
-    return group;
-
-};
-
-var create_extra_user = async function(user_name, group, internal_user){
-    let password = Math.random().toString(36).slice(-10);
-    if(process.env.MY_ADMIN_PASSWORD){
-        password = process.env.MY_ADMIN_PASSWORD;
-    }
-    else {
-        logger.info('Generated admin password:' + password);
-    }
-
-    let user = {
-        status: STATUS_ACTIVE,
-        uid: user_name,
-        firstname: user_name,
-        lastname: user_name,
-        email: process.env.MY_ADMIN_EMAIL || CONFIG.general.support,
-        address: '',
-        lab: '',
-        responsible: '',
-        group: (CONFIG.general.disable_user_group) ? '' : group.name,
-        secondarygroups: [],
-        maingroup: '',
-        home: '',
-        why: '',
-        ip: '',
-        regkey: Math.random().toString(36).slice(-10),
-        is_internal: internal_user,
-        is_fake: false,
-        uidnumber: -1,
-        gidnumber: -1,
-        cloud: false,
-        duration: '1 year',
-        expiration: new Date().getTime() + day_time*360,
-        loginShell: '/bin/bash',
-        history: [],
-        password: password
-    };
-    let minuid = await utils.getUserAvailableId();
-    user.uidnumber = minuid;
-    user.gidnumber = (CONFIG.general.disable_user_group) ? -1 : group.gid;
-    user.home = get_user_home(user);
-    let fid = new Date().getTime();
-    try {
-        await goldap.add(user, fid);
-    } catch(err) {
-        logger.error('Failed to create extra user', user, err);
-        throw err;
-    }
-    logger.debug('user added to ldap');
-
-    delete user.password;
-    // eslint-disable-next-line no-unused-vars
-    await utils.mongo_users().insertOne(user);
-    try {
-        let created_file = await filer.user_create_extra_user(user, fid);
-        logger.info('File Created: ', created_file);
-    } catch(error){
-        logger.error('Create User Failed for: ' + user.uid, error);
-        throw error;
-    }
-
-    let plugin_call = function(plugin_info, userId, data, adminId){
-        // eslint-disable-next-line no-unused-vars
-        return new Promise(function (resolve, reject){
-            let plugins_modules = utils.plugins_modules();
-            plugins_modules[plugin_info.name].activate(userId, data, adminId).then(function(){
-                resolve(true);
-            });
-        });
-    };
-
-    try {
-        let plugins_info = utils.plugins_info();
-        await Promise.all(plugins_info.map(function(plugin_info){
-            return plugin_call(plugin_info, user.uid, user, 'auto');
-        }));
-    } catch(err) {
-        logger.error('failed to create extra user', user, err);
-    }
-    return user;
-};
-
-router.create_admin = async function(default_admin, default_admin_group){
-    let user = await utils.mongo_users().findOne({'uid': default_admin});
-    if(user){
-        logger.info('admin already exists, skipping');
-    }
-    else {
-        logger.info('should create admin');
-        let group = await utils.mongo_groups().findOne({name: default_admin_group});
-        if(group){
-            logger.info('group already exists');
-            // eslint-disable-next-line no-unused-vars
-            await create_extra_user(default_admin, group, true);
-            logger.info('admin user created');
-        }
-        else {
-            logger.info('need to create admin group');
-            let group = await create_group(default_admin_group, default_admin);
-            logger.info('admin group created', group);
-            let user = await create_extra_user(default_admin, group, true);
-            logger.info('admin user created', user);
-        }
-    }
-};
 
 router.get('/user/:id/apikey', async function(req, res){
     if(! req.locals.logInfo.is_logged) {
@@ -415,38 +264,6 @@ router.get('/group/:id', async function(req, res){
     res.end();
 });
 
-router.delete_group = async function(group, admin_user_id){
-    await utils.mongo_groups().deleteOne({'name': group.name});
-    let fid = new Date().getTime();
-    await goldap.delete_group(group, fid);
-    try {
-        let created_file = await filer.user_delete_group(group, fid);
-        logger.info('File Created: ', created_file);
-    } catch(error){
-        logger.error('Delete Group Failed for: ' + group.name, error);
-        return false;
-    }
-
-    await utils.mongo_events().insertOne({'owner': admin_user_id, 'date': new Date().getTime(), 'action': 'delete group ' + group.name , 'logs': [group.name + '.' + fid + '.update']});
-    await utils.freeGroupId(group.gid);
-    return true;
-};
-
-router.clear_user_groups = async function(user, admin_user_id){
-    let allgroups = user.secondarygroups;
-    if (user.group && user.group != '') {
-        allgroups.push(user.group);
-    }
-    for(let i=0;i < allgroups.length;i++){
-        let group = await utils.mongo_groups().findOne({name: allgroups[i]});
-        if(group){
-            let users_in_group = await utils.mongo_users().find({'$or': [{'secondarygroups': group.name}, {'group': group.name}]}).toArray();
-            if(users_in_group && users_in_group.length == 0){
-                router.delete_group(group, admin_user_id);
-            }
-        }
-    }
-};
 
 
 router.delete('/group/:id', async function(req, res){
@@ -477,7 +294,7 @@ router.delete('/group/:id', async function(req, res){
         res.status(403).send({message: 'Group has some users, cannot delete it'});
         return;
     }
-    router.delete_group(group, user.uid).then(function(){
+    grpsrv.delete_group(group, user.uid).then(function(){
         res.send({message: 'group ' + req.params.id + ' deleted'});
         res.end();
     });
@@ -659,44 +476,6 @@ router.get('/user', async function(req, res) {
     res.json(users);
 });
 
-var add_user_to_group = async function (uid, secgroup, action_owner) {
-    logger.info('Adding user ' + uid + ' to group ' + secgroup);
-    let user = await utils.mongo_users().findOne({uid: uid});
-    if(!user){
-        throw {code: 404, message:'User not found'};
-    }
-    if(secgroup == user.group) {
-        throw {code: 208, message: 'Group is user main\'s group: '+user.group};
-    }
-    if(!user.secondarygroups) {
-        user.secondarygroups = [];
-    }
-    for(let g=0;g < user.secondarygroups.length;g++){
-        if(secgroup == user.secondarygroups[g]) {
-            throw {code: 208, message: 'group is already set'};
-        }
-    }
-    user.secondarygroups.push(secgroup);
-    let fid = new Date().getTime();
-    // Now add group
-    await goldap.change_user_groups(user, [secgroup], [], fid);
-    try {
-        await utils.mongo_users().updateOne({_id: user._id}, {'$set': { secondarygroups: user.secondarygroups}});
-    } catch(err) {
-        throw {code: 500, message: 'Could not update user'};
-    }
-
-    try {
-        let created_file = await filer.user_change_group(user, fid);
-        logger.info('File Created: ', created_file);
-        await utils.mongo_events().insertOne({'owner': action_owner, 'date': new Date().getTime(), 'action': 'add user ' + uid + ' to secondary  group ' + secgroup , 'logs': [created_file]});
-    } catch(error){
-        logger.error('Group Change Failed for: ' + user.uid, error);
-        throw {code: 500, message:'Change Group Failed'};
-    }
-
-
-};
 
 router.post('/user/:id/group/:group', async function(req, res){
     if(! req.locals.logInfo.is_logged) {
@@ -721,7 +500,7 @@ router.post('/user/:id/group/:group', async function(req, res){
     let secgroup = req.params.group;
 
     try {
-        await add_user_to_group(uid, secgroup, session_user.uid);
+        await usrsrv.add_user_to_group(uid, secgroup, session_user.uid);
     } catch (e) {
         logger.error(e);
         if (e.code && e.message) {
@@ -813,97 +592,13 @@ router.delete('/user/:id/group/:group', async function(req, res){
         res.end();
         return;
     }
-    router.delete_group(group, session_user.uid).then(function(){
+    grpsrv.delete_group(group, session_user.uid).then(function(){
         res.send({message: 'User removed from group. Empty group ' + secgroup + ' was deleted'});
         res.end();
         return;
     });
 });
 
-router.delete_user = async function(user, action_owner_id, message){
-    let user_is_activ = true;
-
-    if(user.status == STATUS_PENDING_EMAIL || user.status == STATUS_PENDING_APPROVAL){
-        user_is_activ = false;
-    }
-
-    let fid = new Date().getTime();
-    // Remove user from groups
-    let allgroups = user.secondarygroups;
-    if (user.group && user.group != '') {
-        allgroups.push(user.group);
-    }
-    if(user_is_activ){
-        await goldap.change_user_groups(user, [], allgroups, fid);
-
-        try {
-            let created_file = await filer.user_delete_user(user, fid);
-            logger.info('File Created: ', created_file);
-        } catch(error){
-            logger.error('Delete User Failed for: ' + user.uid, error);
-            return false;
-        }
-    }
-
-    try {
-        await utils.mongo_users().deleteOne({_id: user._id});
-    } catch(err) {
-        return false;
-    }
-
-    router.clear_user_groups(user, action_owner_id);
-
-    await utils.mongo_events().insertOne({
-        'owner': action_owner_id,
-        'date': new Date().getTime(),
-        'action': 'delete user ' + user.uid ,
-        'logs': [user.uid + '.' + fid + '.update']
-    });
-
-    let msg_destinations =  [GENERAL_CONFIG.accounts];
-    let mail_message=  'no explaination provided !';
-    if (message && message.length > 1) {
-        mail_message = message;
-        msg_destinations.push(user.email);
-    }
-
-    try {
-        await utils.send_notif_mail({
-            'name': 'user_deletion',
-            'destinations': msg_destinations,
-            'subject': 'account deletion: ' + user.uid
-        }, {
-            '#UID#': user.uid,
-            '#USER#': action_owner_id,
-            '#MSG#': mail_message
-        });
-    } catch(error) {
-        logger.error(error);
-    }
-
-    if(user_is_activ){
-        // Call remove method of plugins if defined
-        let plugin_call = function(plugin_info, userId, user, adminId){
-            // eslint-disable-next-line no-unused-vars
-            return new Promise(function (resolve, reject){
-                let plugins_modules = utils.plugins_modules();
-                if(plugins_modules[plugin_info.name].remove === undefined) {
-                    resolve(true);
-                }
-                plugins_modules[plugin_info.name].remove(userId, user, adminId).then(function(){
-                    resolve(true);
-                });
-            });
-        };
-        let plugins_info = utils.plugins_info();
-        await Promise.all(plugins_info.map(function(plugin_info){
-            return plugin_call(plugin_info, user.uid, user, action_owner_id);
-        }));
-    }
-
-    await utils.freeUserId(user.uidnumber);
-    return true;
-};
 
 router.delete('/user/:id', async function(req, res){
     if(! req.locals.logInfo.is_logged) {
@@ -934,7 +629,7 @@ router.delete('/user/:id', async function(req, res){
         return;
     }
     if(user.status == STATUS_PENDING_EMAIL || user.status == STATUS_PENDING_APPROVAL){
-        router.delete_user(user, session_user.uid, mail_message).then(function(){
+        usrsrv.delete_user(user, session_user.uid, mail_message).then(function(){
             res.send({message: 'User deleted'});
             res.end();
             return;
@@ -956,7 +651,7 @@ router.delete('/user/:id', async function(req, res){
             res.end();
             return;
         }
-        router.delete_user(user, session_user.uid, mail_message).then(function(){
+        usrsrv.delete_user(user, session_user.uid, mail_message).then(function(){
             res.send({message: 'User deleted'});
             res.end();
             return;
@@ -1021,7 +716,7 @@ router.get('/user/:id/activate', async function(req, res) {
         user.gidnumber = data.gid;
     }
 
-    user.home = get_user_home(user);
+    user.home = usrsrv.get_user_home(user);
     let fid = new Date().getTime();
     try {
         await goldap.add(user, fid);
@@ -1308,7 +1003,7 @@ router.post('/user/:id', async function(req, res) {
         history: [{action: 'register', date: new Date().getTime()}]
     };
     // user[CONFIG.general.internal_flag] = false,
-    user.home = get_user_home(user);
+    user.home = usrsrv.get_user_home(user);
 
     await utils.mongo_events().insertOne({'owner': req.params.id, 'date': new Date().getTime(), 'action': 'user registration ' + req.params.id , 'logs': []});
 
@@ -1958,7 +1653,7 @@ router.put('/user/:id', async function(req, res) {
         if (req.body.maingroup) {
             user.maingroup = req.body.maingroup;
         }
-        user.home = get_user_home(user);
+        user.home = usrsrv.get_user_home(user);
 
     }
 
@@ -2020,7 +1715,7 @@ router.put('/user/:id', async function(req, res) {
         if(users_in_group && users_in_group.length == 0){
             let oldgroup = await utils.mongo_groups().findOne({name: user.oldgroup});
             if(oldgroup){
-                router.delete_group(oldgroup, session_user.uid);
+                grpsrv.delete_group(oldgroup, session_user.uid);
             }
         }
 
@@ -2048,7 +1743,7 @@ router.put('/user/:id', async function(req, res) {
         if(users_in_group && users_in_group.length == 0){
             let oldgroup = await utils.mongo_groups().findOne({name: user.oldgroup});
             if(oldgroup){
-                await router.delete_group(oldgroup, session_user.uid);
+                await grpsrv.delete_group(oldgroup, session_user.uid);
             }
         }
         user.fid = null;
@@ -2075,73 +1770,6 @@ router.get('/project/:id/users', async function(req, res){
     res.end();
 });
 
-var add_user_to_project = async function (newproject, uid, action_owner) {
-    logger.info('Adding user ' + uid + ' to project ' + newproject);
-
-    let fid = new Date().getTime();
-    let user = await utils.mongo_users().findOne({uid: uid});
-    if(!user) {
-        throw {code: 404, message: 'User does not exist'};
-    }
-    if (!user.projects){
-        user.projects = [];
-    }
-    for(let g=0; g < user.projects.length; g++){
-        if(newproject == user.projects[g]) {
-            throw {code: 208, message: 'User is already in project : nothing was done.'};
-        }
-    }
-    user.projects.push(newproject);
-    try {
-        await utils.mongo_users().updateOne({_id: user._id}, {'$set': { projects: user.projects}});
-    } catch(err) {
-        throw {code: 500, message: 'Could not update user'};
-    }
-
-    try {
-        let created_file = await filer.project_add_user_to_project({id: newproject}, user, fid);
-        logger.info('File Created: ', created_file);
-        await utils.mongo_events().insertOne({'owner': action_owner, 'date': new Date().getTime(), 'action': 'add user ' + uid + ' to project ' + newproject , 'logs': [created_file]});
-    } catch(error){
-        logger.error(error);
-        throw {code: 500, message:'Add User to Project Failed for: ' + newproject};
-    }
-
-    let project = await utils.mongo_projects().findOne({id:newproject});
-    let msg_destinations = [user.email];
-    let owner = await utils.mongo_users().findOne({uid:project.owner});
-    if (owner) {
-        msg_destinations.push(owner.email);
-    }
-
-    try {
-        await utils.send_notif_mail({
-            'name': 'add_to_project',
-            'destinations': msg_destinations,
-            'subject': 'account ' + user.uid + ' added to project : ' + project.id
-        }, {
-            '#UID#': user.uid,
-            '#NAME#': project.id,
-            '#SIZE#': project.size,
-            '#DESC#': project.description,
-            '#PATH#': project.path
-        });
-    } catch(error) {
-        logger.error(error);
-    }
-
-    if (CONFIG.project === undefined || CONFIG.project.enable_group) {
-        try {
-            await add_user_to_group(user.uid, project.group, action_owner);
-        } catch (error) {
-            logger.error(error);
-            // as it may throw any 20* http ok code...
-            if (!error.code || error.code >= 300) {
-                throw {code: 500, message:'Add User to group;' + project.group + ' Failed for project: ' + project.id};
-            }
-        }
-    }
-};
 
 router.post('/user/:id/project/:project', async function(req, res){
     if(! req.locals.logInfo.is_logged) {
@@ -2164,7 +1792,7 @@ router.post('/user/:id/project/:project', async function(req, res){
     let uid = req.params.id;
 
     try {
-        await add_user_to_project(newproject, uid, session_user.uid);
+        await usrsrv.add_user_to_project(newproject, uid, session_user.uid);
     } catch (e) {
         logger.error(e);
         if (e.code && e.message) {
